# docs/benchmarks/schemas/openapi.yaml
...
openapi: 3.1.0
info:
  title: AgentLayer Admin — Agent Config & Benchmark Tuning
  version: 0.1.0-planned
  description: |
    Planned endpoints under `/v1/admin/agent-config` and extensions to `/v1/admin/benchmarks`.
    Existing operator-settings and benchmark run APIs are referenced but not fully specified here.
...
paths:
  /v1/admin/agent-config/knobs:
    get:
      tags: [agent-config]
      operationId: agentConfigListKnobs
      summary: List knobs from registry with effective values
      parameters:
        - $ref: '#/components/parameters/UiGroupQuery'
        - $ref: '#/components/parameters/LayerQuery'
        - name: benchmark_sensitive
          in: query
          schema: { type: boolean }
        - name: agent_id
          in: query
          schema: { type: string }
        - name: writable_only
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Knob list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnobsListResponse' }
  /v1/admin/agent-config/knobs/{knob_id}:
    get:
      tags: [agent-config]
      operationId: agentConfigGetKnob
      summary: Single knob detail
      parameters:
        - $ref: '#/components/parameters/KnobIdPath'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KnobDetailResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/admin/agent-config/changelog:
    get:
      tags: [agent-config]
      operationId: agentConfigGetChangelog
      summary: Paginated config change audit log
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: experiment_id
          in: query
          schema: { type: string, format: uuid }
        - name: session_id
          in: query
          schema: { type: string, format: uuid }
        - name: actor_type
          in: query
          schema: { $ref: '#/components/schemas/ActorType' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangelogListResponse' }
  /v1/admin/agent-config/draft:
    post:
      tags: [agent-config]
      operationId: agentConfigCreateDraft
      summary: Validate patches without applying
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DraftRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DraftResponse' }
        '400':
          $ref: '#/components/responses/ValidationError'
  /v1/admin/agent-config/apply:
    post:
      tags: [agent-config]
      operationId: agentConfigApply
      summary: Apply runtime knob patches (primary tuning write path)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplyRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplyResponse' }
        '400':
          $ref: '#/components/responses/ValidationError'
  /v1/admin/agent-config/sessions:
    get:
      tags: [agent-config-sessions]
      operationId: agentConfigListSessions
      parameters:
        - $ref: '#/components/parameters/LimitQuery'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/SessionStatus' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionListResponse' }
    post:
      tags: [agent-config-sessions]
      operationId: agentConfigCreateSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionCreateRequest' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
  /v1/admin/agent-config/sessions/{session_id}:
    get:
      tags: [agent-config-sessions]
      operationId: agentConfigGetSession
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [agent-config-sessions]
      operationId: agentConfigPatchSession
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionPatchRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
  /v1/admin/agent-config/sessions/{session_id}/validate:
    post:
      tags: [agent-config-sessions]
      operationId: agentConfigValidateSession
      summary: Queue validation benchmark for session
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionValidateRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionValidateResponse' }
  /v1/admin/agent-config/sessions/{session_id}/close:
    post:
      tags: [agent-config-sessions]
      operationId: agentConfigCloseSession
      parameters:
        - $ref: '#/components/parameters/SessionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionCloseRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
...
components:
  parameters:
    KnobIdPath:
      name: knob_id
      in: path
      required: true
      schema: { type: string, example: agent.max_tool_rounds }
    SessionIdPath:
      name: session_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
    UiGroupQuery:
      name: ui_group
      in: query
      schema: { type: string }
    LayerQuery:
      name: layer
      in: query
      schema: { $ref: '#/components/schemas/KnobLayer' }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    OkWrapper:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, const: true }
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
    KnobLayer:
      type: string
      enum: [operator, runtime_config, agent_yaml, router_yaml, bench, rubric, code]
    KnobSource:
      type: string
      enum: [db_override, bootstrap_install_default, file_default, operator_settings, git]
    KnobValueType:
      type: string
      enum: [boolean, integer, number, string, string_list, json]
    UiGroup:
      type: object
      required: [id, label]
      properties:
        id: { type: string }
        label: { type: string }
    KnobDescriptor:
      type: object
      required: [id, layer, ui_group, writable, benchmark_sensitive]
      properties:
        id: { type: string, example: agent.max_tool_rounds }
        layer: { $ref: '#/components/schemas/KnobLayer' }
        ui_group: { type: string }
        writable: { type: boolean }
        type: { $ref: '#/components/schemas/KnobValueType' }
        default: {}
        effective: {}
        source: { $ref: '#/components/schemas/KnobSource' }
        bootstrap_env_key: { type: string, nullable: true, description: Install default only — never tuned via .env UI }
        source_path: { type: string, nullable: true }
        content_hash: { type: string, nullable: true, description: sha256 for non-writable layers }
        affects_agents:
          type: array
          items: { type: string }
        affects_clusters:
          type: array
          items: { type: string }
        benchmark_sensitive: { type: boolean }
        doc: { type: string }
    KnobsListResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [registry_version, ui_groups, knobs]
          properties:
            registry_version: { type: integer }
            ui_groups:
              type: array
              items: { $ref: '#/components/schemas/UiGroup' }
            knobs:
              type: array
              items: { $ref: '#/components/schemas/KnobDescriptor' }
    KnobDetailResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [knob]
          properties:
            knob: { $ref: '#/components/schemas/KnobDescriptor' }
            last_changelog_event: { $ref: '#/components/schemas/ChangelogEvent', nullable: true }
...
    ActorType:
      type: string
      enum: [user, operator_agent, reviewer_agent, reviewer_job, system]
    Actor:
      type: object
      required: [type]
      properties:
        type: { $ref: '#/components/schemas/ActorType' }
        user_id: { type: string, format: uuid, nullable: true }
        agent_id: { type: string, nullable: true }
        conversation_id: { type: string, format: uuid, nullable: true }
    KnobPatch:
      type: object
      required: [knob_id, value]
      properties:
        knob_id: { type: string }
        value: {}
    AppliedPatch:
      type: object
      required: [knob_id, old, new]
      properties:
        knob_id: { type: string }
        old: {}
        new: {}
    SkippedPatch:
      type: object
      required: [knob_id, reason]
      properties:
        knob_id: { type: string }
        reason:
          type: string
          enum: [not_writable, validation_failed, unchanged, unknown_knob]
    ChangelogEvent:
      type: object
      required: [id, at, actor, patches, fingerprint_before, fingerprint_after]
      properties:
        id: { type: string, format: uuid }
        at: { type: string, format: date-time }
        actor: { $ref: '#/components/schemas/Actor' }
        session_id: { type: string, format: uuid, nullable: true }
        experiment_id: { type: string, format: uuid, nullable: true }
        hypothesis: { type: string, nullable: true }
        patches:
          type: array
          items: { $ref: '#/components/schemas/AppliedPatch' }
        fingerprint_before: { type: string }
        fingerprint_after: { type: string }
    ChangelogListResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [events]
          properties:
            events:
              type: array
              items: { $ref: '#/components/schemas/ChangelogEvent' }
    BenchmarkTrigger:
      type: object
      properties:
        suite_preset:
          type: string
          enum: [routing-core, smoke, full, custom]
        harness_preset:
          type: string
          enum: [observability, chat_parity]
        profiles_from_saved_matrix: { type: boolean, default: true }
        suite: { type: string, description: Required when suite_preset is custom }
        scenarios:
          type: array
          items: { type: string }
    DraftRequest:
      type: object
      required: [patches]
      properties:
        patches:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/KnobPatch' }
        hypothesis: { type: string, maxLength: 4000 }
    DraftValidation:
      type: object
      required: [valid]
      properties:
        valid: { type: boolean }
        fingerprint_preview: { type: string, nullable: true }
        errors:
          type: array
          items:
            type: object
            properties:
              knob_id: { type: string }
              message: { type: string }
        warnings:
          type: array
          items: { type: string }
    DraftResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [draft_id, validation]
          properties:
            draft_id: { type: string, format: uuid }
            validation: { $ref: '#/components/schemas/DraftValidation' }
    ApplyRequest:
      type: object
      required: [patches]
      properties:
        patches:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/KnobPatch' }
        session_id: { type: string, format: uuid, nullable: true }
        experiment_id: { type: string, format: uuid, nullable: true }
        hypothesis: { type: string, maxLength: 4000, nullable: true }
        trigger_benchmark: { type: boolean, default: false }
        benchmark: { $ref: '#/components/schemas/BenchmarkTrigger' }
        actor:
          description: Set by server from auth; clients omit
          readOnly: true
          $ref: '#/components/schemas/Actor'
    ApplyResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [applied, skipped, fingerprint, changelog_event_id]
          properties:
            applied:
              type: array
              items: { $ref: '#/components/schemas/AppliedPatch' }
            skipped:
              type: array
              items: { $ref: '#/components/schemas/SkippedPatch' }
            fingerprint: { type: string }
            changelog_event_id: { type: string, format: uuid }
            benchmark_run_id: { type: string, format: uuid, nullable: true }
    SessionStatus:
      type: string
      enum: [open, validating, closed_accepted, closed_rejected, closed_reverted]
    TuningSession:
      type: object
      required: [id, status, baseline_fingerprint, cohort_label, created_at]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        hypothesis: { type: string, nullable: true }
        cohort_label: { type: string }
        status: { $ref: '#/components/schemas/SessionStatus' }
        baseline_fingerprint: { type: string }
        current_fingerprint: { type: string, nullable: true }
        experiment_ids:
          type: array
          items: { type: string, format: uuid }
        run_ids:
          type: array
          items: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        closed_at: { type: string, format: date-time, nullable: true }
    SessionCreateRequest:
      type: object
      required: [cohort_label]
      properties:
        label: { type: string, maxLength: 128 }
        hypothesis: { type: string, maxLength: 4000 }
        cohort_label: { type: string, maxLength: 128 }
    SessionPatchRequest:
      type: object
      properties:
        label: { type: string }
        hypothesis: { type: string }
    SessionValidateRequest:
      type: object
      properties:
        suite_preset: { type: string, default: routing-core }
        harness_preset: { type: string, default: observability }
        profiles_from_saved_matrix: { type: boolean, default: true }
    SessionValidateResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [session, benchmark_run_id]
          properties:
            session: { $ref: '#/components/schemas/TuningSession' }
            benchmark_run_id: { type: string, format: uuid }
    SessionCloseRequest:
      type: object
      required: [accept]
      properties:
        accept: { type: boolean }
        revert_patches: { type: boolean, default: false, description: If accept=false, apply inverse patches }
    SessionResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [session]
          properties:
            session: { $ref: '#/components/schemas/TuningSession' }
    SessionListResponse:
      allOf:
        - $ref: '#/components/schemas/OkWrapper'
        - type: object
          required: [sessions]
          properties:
            sessions:
              type: array
              items: { $ref: '#/components/schemas/TuningSession' }
...